Deleted the old login page folder and moved all relevant code to the main page, also changed the register page so that it would point to the new login page, but other pages will need to be updated as well
