+++ with-mongodb-app/app/layout.tsx
@@ -1,27 +1,12 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "@/styles/globals.css";
-import {
-  SidebarInset,
-  SidebarProvider,
-  SidebarTrigger,
-} from "@/components/ui/sidebar";
-import { AppSidebar } from "@/components/app-sidebar";
-import {
-  Breadcrumb,
-  BreadcrumbList,
-  BreadcrumbItem,
-  BreadcrumbLink,
-  BreadcrumbSeparator,
-  BreadcrumbPage,
-} from "@/components/ui/breadcrumb";
-import { Separator } from "@radix-ui/react-separator";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Login - Hobby Helper",
+  description: "The Login Page for the Hobby Helper webapp",
 };
 
 export default function RootLayout({
@@ -30,9 +15,8 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <>
-    <html><body>{children}</body></html>
-    </>
-    
+    <html>
+      <body>{children}</body>
+    </html>
   );
 }
