--- with-mongodb-app/app/login/layout.tsx
@@ -1,22 +1,0 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "@/styles/globals.css";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Login - Hobby Helper",
-  description: "The Login Page for the Hobby Helper webapp",
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html>
-      <body>{children}</body>
-    </html>
-  );
-}
